Fix DB memory leak in transaction containers
Transactions are stored in the DB via a multi-index container for purposes of deduplication checks. The index based on expiry is ordered_unique which means items will be sorted ascending for that index. So when pruning old transactions from this container we need to be removing from the start of the index as that's where the oldest transactions will be.

Issue #2153