fix compiler_builtins long double comparisons to fix test for #1954
Clang was actually generating the right code; compiler_builtins had the problem because the return value semantics of the Berkeley softfloat library are not the same as that of the compiler builtin intrinsics.

My fixes will work for now until we replace the compiler_builtins intrinsics altogether with the change described in #2487.