contracts are now only granted a virtual permission of "eosio.code"
A contract no longer satisfies all permissions of its account by default.

If a contract needs to be able send an inline action or deferred transaction and requires adding some specific permission to the authorizations of the action, then the authority of that permission must be set up in such a way that it can be satisfied by just the "eosio.code" permission.