Add the concept of persisted transactions to the producer plugin.
persisted transactions are the transactions _directly submitted to that node_.  They should only exist on RPC nodes that are accepting transactions from users.  As such the expectation (from our smoke tests at least) is that said node will persist the state of the chain + that transaction until either the block-chain represents the side effects of the transaction OR the transaction expires.

From the pov of the producer plugin, this means that even when speculating on a block the block should include these transactions.