Fix message_buffer peak; make peak & read throw on error
This commit fixes two problems with message_buffer. First, users of read() (like fc::raw::unpack) expect that on an out of bounds read that read() throws. You can see this behavior in fc::datastream where its read() will throw an out_of_range_exception. The boolean returned from fc:: datastream::read() is always true. So for both message_buffer read() and peak() throw an out_of_range_exception when attempting to read or peak beyond the end. peak() is something that is only used internally to net_plugin and it appears that net_plugin expects it to throw too as all uses of it don't check for its return value.

The second fix is that peak()'s logic for detecting end of buffer didn't seem right. It always based the number of bytes remaining on the read index no matter where we are peaking from. So that needed fixing too.